Rename darkmode.js to theme.js
The module has only owned the theme system since the terminal engine moved to
terminal.js; the darkmode.js name predates even the palette/typography/layout
work folded into it. Rename the file (and its two test files) to match what it
is, and repoint every reference in shipping code — head.html's resource path
and the test requires (functional), plus the comments across the layout
partials, terminal.css, the sibling JS modules, and the tests.

Comments that described terminal behaviour now point at terminal.js (the boot
theatre, cwd/chdir, slug stamping, cd echo, keybar, flows, colophon); theme
behaviour points at theme.js (layout pairings, coty transport, transitions).
AGENTS.md's naming examples updated. Dated planning/analysis docs keep the old
name as historical record. Full suite green (626).
