+++ assets/js/__tests__/terminal-commands.test.js
@@ -1,5 +1,5 @@
 /**
- * Terminal command line (darkmode.js): the tail <input> that fixes mobile
+ * Terminal command line (terminal.js): the tail <input> that fixes mobile
  * typing and hosts the small command set + easter eggs. Driven the same way as
  * darkmode-pantone.test.js — require the module, fire DOMContentLoaded, then
  * dispatch real events at the input.
@@ -30,9 +30,9 @@ describe("terminal command line", () => {
 
   function loadModule() {
     jest.isolateModules(() => {
-      // darkmode.js first (publishes window.Theme synchronously), then the
+      // theme.js first (publishes window.Theme synchronously), then the
       // terminal engine, which aliases window.Theme at IIFE eval.
-      require("../darkmode.js");
+      require("../theme.js");
       require("../terminal.js");
     });
     document.dispatchEvent(new window.Event("DOMContentLoaded"));
@@ -811,7 +811,7 @@ describe("terminal command line", () => {
     loadModule();
     typeCommand("konami");
     expect(document.documentElement.getAttribute("data-konami")).toBe("on");
-    // toggleKonami de-dupes a synchronous burst (see darkmode.js), so let the
+    // toggleKonami de-dupes a synchronous burst (see terminal.js), so let the
     // guard clear on the microtask queue before toggling back off.
     await Promise.resolve();
     typeCommand("konami");
