+++ layouts/_default/legal.html
@@ -16,7 +16,7 @@
   {{ if .IsSection }}
     {{/* The legal/ section index lists its pages as cards — plain links in
       most layouts, harvested as `license.md`/`privacy.md` files in the
-      terminal (see darkmode.js's remote-ls).
+      terminal (see terminal.js's remote-ls).
     */}}
     {{ range .Pages }}
       <article class="summary-card place-prose">
